Did assignment 0

+++ 03_code2/code2.c
@@ -3,7 +3,7 @@
 int printTriangle(int size) {
   //start with starCount being 0
 int  i,j;
-  int starCount;
+float starCount;
     //count from 0 (inclusive) to size (exclusive), for each number i that you count
   for (i=0; i<size; i++){
      //count from 0 (inclusive) to i (inclusive), for each number j that you count
